Catch up main: 3 weeks of work
Brings origin/main current with local. Notable adds:

- Lotto v2 momentum/regime gates (G2/G3) + G4 trigger-bar rule
- New `src/index_swing/` module (Tier 4 swing skill)
- New `src/recovery_plan/` module — zero-default state for new
  installs; user fills in via dashboard
- Unified verdict surface via `src/action_gate/` + `src/scan_verdict`
- Sunday-scan UI/route retirement (covered by dashboard now)
- Backtest harnesses for lotto cohort, free-range, index-swing TF,
  apex options, capex detection, spread-haircut analysis
- New views: IndexSwing, Recovery, DailyWisdom, TradeCard
- Polygon loader (paused, kept for future re-enable)
- 30+ new test files covering the new modules

.gitignore: exclude scripts/*.csv|json (per-user analysis output)
and frontend build artifacts.

+++ src/positions/cli.py
@@ -188,7 +188,7 @@ def build_parser() -> argparse.ArgumentParser:
     po.add_argument("--expiry")
     po.add_argument("--premium", type=float, help="Premium per share for options")
     po.add_argument("--contracts", type=int)
-    po.add_argument("--shares", type=int)
+    po.add_argument("--shares", type=float)
     po.add_argument("--entry-price", type=float,
                     help="Underlying price at entry (options) or fill price (shares)")
     po.add_argument("--target", type=float, help="Target price for the underlying")
